requirements.txt
aniso8601==9.0.1
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.6.0
certifi==2023.5.7
chardet==3.0.4
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
finance-datareader==0.9.50
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.39.3
gensim==3.8.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.0.0
joblib==0.15.1
JPype1==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
kochat==1.0.3
kocrawl==1.0.9
konlpy==0.5.2
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.2.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.18.5
oauthlib==3.2.2
OpenDartReader==0.2.3
packaging==23.1
pandas==1.0.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.3
pytorch-crf==0.7.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.1
scikit-learn==0.23.1
scipy==1.9.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
stack-data==0.6.2
threadpoolctl==3.1.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tweepy==3.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.25.11
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0
